refactor: 重构API Key管理器文件命名规范
- 将api-key-manager.ts重命名为api_key_manager.ts以符合项目命名规范
- 更新相关导入路径和引用
- 保持代码功能不变，仅调整文件命名一致性
